Add files via upload
In this project, a rectangular city area is surrounded by 4 different cities on each of its border. The rectangular city area is comprised of a grid system where the altitude of different nodes is stored using Python program with the help of graph data structures. The goal is to use graph data structure and perform BFS or DFS to find the route between 2 nodes and the route between 2 cities. Here we are using 2D matrix list to store the data in graph.In this project, a rectangular city area is surrounded by 4 different cities on each of its border. The rectangular city area is comprised of a grid system where the altitude of different nodes is stored using Python program with the help of graph data structures. The goal is to use graph data structure and perform BFS or DFS to find the route between 2 nodes and the route between 2 cities. Here we are using 2D matrix list to store the data in graph.
Skills: 2D matrix · searching techniques